# Conflicts:
#	MiniRPG/Assets/MinaMapCamera.cs.meta
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem.meta
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/CollectCoinsQuestStep.cs
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/CollectCoinsQuestStep.cs.meta
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/Quest.cs
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/Quest.cs.meta
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/QuestInfoSO.cs
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/QuestInfoSO.cs.meta
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/QuestManager.cs
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/QuestManager.cs.meta
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/QuestStep.cs
#	MiniRPG/Assets/Scripts/Core/System/QuestSystem/QuestStep.cs.meta
#	MiniRPG/Assets/Scripts/Managers/EntityManager.cs
#	MiniRPG/Assets/Scripts/Managers/EntityManager.cs.meta
#	MiniRPG/Assets/Scripts/Managers/QuestManager.cs.meta
#	MiniRPG/Assets/Scripts/Quest.meta
#	MiniRPG/Assets/Scripts/Quest/Quest.cs
#	MiniRPG/Assets/Scripts/Quest/Quest.cs.meta
#	MiniRPG/Assets/Scripts/QuestSystem.meta
#	MiniRPG/Assets/Scripts/QuestSystem/Quest.cs
#	MiniRPG/Assets/Scripts/QuestSystem/Quest.cs.meta
#	MiniRPG/Assets/Scripts/QuestSystem/QuestInfoSO.cs.meta
#	MiniRPG/Assets/Scripts/QuestSystem/QuestManager.cs.meta

+++ MiniRPG/Assets/Scripts/UI/Scene/MainSceneUI.cs
@@ -55,15 +55,15 @@ public class MainSceneUI : BaseUI
 
         _hpText.text = ((int)(_player.PlayerData.Hp.CurValue)).ToString();
         _hpMaxText.text = ((int)(_player.PlayerData.Hp.MaxValue)).ToString();
-        _hpBar.fillAmount = (int)(_player.PlayerData.Hp.CurValue) / (int)(_player.PlayerData.Hp.MaxValue);
+        _hpBar.fillAmount = (_player.PlayerData.Hp.CurValue) / (_player.PlayerData.Hp.MaxValue);
 
         _mpText.text = ((int)(_player.PlayerData.Mp.CurValue)).ToString();
         _mpMaxText.text = ((int)(_player.PlayerData.Mp.MaxValue)).ToString();
-        _mpBar.fillAmount = (int)(_player.PlayerData.Mp.CurValue) / (int)(_player.PlayerData.Mp.MaxValue);
+        _mpBar.fillAmount = (_player.PlayerData.Mp.CurValue) / (_player.PlayerData.Mp.MaxValue);
 
         _expText.text = ((int)(_player.PlayerData.Exp.CurValue)).ToString();
         _expMaxText.text = ((int)(_player.PlayerData.Exp.MaxValue)).ToString();
-        _expBar.fillAmount = (int)(_player.PlayerData.Exp.CurValue) / (int)(_player.PlayerData.Exp.MaxValue);
+        _expBar.fillAmount = (_player.PlayerData.Exp.CurValue) / (_player.PlayerData.Exp.MaxValue);
     }
 
     public void AddQuestList(List<Quest> quests)
